requirements.txt
requests
PyYAML
beautifulsoup4
lxml
